requirements.txt
azureml-sdk
azureml-widgets
azureml-mlflow
azure-ai-ml
black